+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
+- **GM-16** - Component Mult Range Slider.
+- **GM-114** - Page plans Gallery filter.
 
 - **GM-113** - ExplanationPlan and Subtitle component.
 ### Fixed
@@ -16,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-**GM-123** - Created Created Banner component on for companies.
+
+- **GM-125** - Created Benefits Plans component business.
 - **GM-121** - Created Benefits Plans component.
 - **GM-297** - Agnostic component for simple text [page plans bottom text].
 - **GM-95** - Created ScrollToTopButton component.
@@ -31,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Fixed
+
 - Removed duplicated TopSellers shelf from home
 
 ## [0.7.0] - 2022-07-04